requirements.txt
requests==2.32.3
google-generativeai==0.7.0
python-dotenv==1.0.1